refactor(compose): give capture's jobs and the tier gate their own files
Two craft findings on files this arc grew, both the same shape — a file that
had become the home of everything near it.

jobs.go crossed the 500-line cap as capture's jobs were added to it. The Gmail
dispatch pass, the per-connection sync it fans out to, and the push-watch
renewal are one concept — how captured mail gets pulled on a schedule — and the
runner file is the composition of every job, not the home of any. They move to
jobs_capture.go, beside the jobs_overlay.go precedent. 500 lines to 369.

The capture suite's first split left the file over the cap and the tier group
still at 109 body lines, so the tiers move to their own file and split by what
they decide: which senders are suppressed, and how correspondence outranks
suppression. Five tests across two files, each named for what it proves.

Backend craft advisories 115 to 113.
