+++ backend/internal/compose/integration/capture_autocreate_integration_test.go
@@ -277,122 +277,6 @@ func TestCaptureAutoCreatesTheCounterpartyBehindAThread(t *testing.T) {
 	})
 }
 
-// The tiered creation gate (ADR-0072 §1), one tier per subtest: which senders
-// become records, which are suppressed, and which precedence decides it.
-func TestCaptureTierGateDecidesWhoBecomesARecord(t *testing.T) {
-	env := newCaptureEnv(t)
-	e, sync, syncSent := env.e, env.sync, env.syncSent
-	t.Run("free-mail creates the person, never a company", func(t *testing.T) {
-		sync(t, email("bob@gmail.com", "Bob Person", autoCreateOwner, "b1@gmail.com", ""))
-		if n := countRows(t, e, `
-			SELECT count(*) FROM person p JOIN person_email pe ON pe.person_id = p.id
-			WHERE pe.email = 'bob@gmail.com'`); n != 1 {
-			t.Fatalf("%d persons for bob, want 1", n)
-		}
-		if n := countRows(t, e, `SELECT count(*) FROM organization WHERE display_name = 'gmail.com'`); n != 0 {
-			t.Fatal("gmail.com must never become an organization")
-		}
-	})
-	t.Run("a corresponded-with free-mail address is still never a company", func(t *testing.T) {
-		// T1 overrides T2 suppression ONLY. Free-mail's org rule is about what a
-		// domain can honestly name, not about whether its sender is trusted, so
-		// writing to a gmail.com address buys its owner a person and never an
-		// organization called "Gmail" — the junk this ADR exists to prevent.
-		syncSent(t, map[string]bool{"fm1@myco.example": true},
-			email(autoCreateOwner, "", "carol@gmail.com", "fm1@myco.example", ""))
-		if n := countRows(t, e, `
-			SELECT count(*) FROM person p JOIN person_email pe ON pe.person_id = p.id
-			WHERE pe.email = 'carol@gmail.com'`); n != 1 {
-			t.Fatalf("%d persons for carol, want 1", n)
-		}
-		if n := countRows(t, e, `SELECT count(*) FROM organization WHERE display_name IN ('Gmail', 'gmail.com')`); n != 0 {
-			t.Fatal("a corresponded-with free-mail address minted an organization")
-		}
-	})
-	t.Run("transactional infrastructure keeps the activity, derives no counterparty", func(t *testing.T) {
-		// A DocuSign envelope (exact infra eSLD, no corroboration needed) and a
-		// conference blast on a prefix subdomain WITH a List-Unsubscribe header
-		// (corroborated) both suppress person+org while the timeline row stands
-		// (ADR-0072/A118, CAP-PARAM-6).
-		sync(t,
-			email("dse@eu.docusign.net", "DocuSign EU", autoCreateOwner, "ds1@docusign.net", ""),
-			emailWithListUnsub("hello@event.gitex.com", "GITEX", autoCreateOwner, "gx1@event.gitex.com"),
-		)
-		if n := countRows(t, e, `SELECT count(*) FROM activity WHERE source_id IN ('ds1@docusign.net', 'gx1@event.gitex.com')`); n != 2 {
-			t.Fatalf("%d transactional activities captured, want 2 — the timeline row must stand", n)
-		}
-		if n := countRows(t, e, `
-			SELECT count(*) FROM person p JOIN person_email pe ON pe.person_id = p.id
-			WHERE pe.email IN ('dse@eu.docusign.net', 'hello@event.gitex.com')`); n != 0 {
-			t.Fatal("transactional infrastructure must derive no person")
-		}
-		if n := countRows(t, e, `SELECT count(*) FROM organization WHERE display_name IN ('Docusign', 'Gitex')`); n != 0 {
-			t.Fatal("transactional infrastructure must derive no organization")
-		}
-		if n := countRows(t, e, `
-			SELECT count(*) FROM system_log
-			WHERE action = 'capture_transactional_suppressed' AND detail->>'source_id' IN ('ds1@docusign.net', 'gx1@event.gitex.com')`); n != 2 {
-			t.Fatalf("%d transactional-suppression breadcrumbs, want 2", n)
-		}
-	})
-	t.Run("writing to an address spares its later bulk mail from suppression", func(t *testing.T) {
-		// T1 runs BEFORE T2 and the order is load-bearing (ADR-0072 §1): once
-		// the workspace has written to someone, their newsletter footer must not
-		// turn them into infrastructure. The same address and the same
-		// List-Unsubscribe corroboration that suppressed above now survive,
-		// because the mailbox owner wrote to them first.
-		syncSent(t, map[string]bool{"ev1@myco.example": true},
-			email(autoCreateOwner, "", "team@event.expo.example", "ev1@myco.example", ""))
-		if n := countRows(t, e, `
-			SELECT count(*) FROM activity
-			WHERE counterparty_email = 'team@event.expo.example' AND counterparty_outbound_attested`); n != 1 {
-			t.Fatalf("%d attested outbound activities, want 1 — the T1 evidence must be stamped", n)
-		}
-
-		sync(t, emailWithListUnsub("team@event.expo.example", "Expo", autoCreateOwner, "ev2@event.expo.example"))
-		if n := countRows(t, e, `
-			SELECT count(*) FROM person p JOIN person_email pe ON pe.person_id = p.id
-			WHERE pe.email = 'team@event.expo.example'`); n != 1 {
-			t.Fatalf("%d persons for a corresponded-with sender, want 1 — T1 must spare it from T2", n)
-		}
-		if n := countRows(t, e, `
-			SELECT count(*) FROM system_log
-			WHERE action = 'capture_correspondence_spared' AND detail->>'source_id' = 'ev2@event.expo.example'`); n != 1 {
-			t.Fatalf("%d spare breadcrumbs, want 1 — an overridden suppression must be as visible as a suppression", n)
-		}
-	})
-	t.Run("a forged From:owner does not whitelist the address it names", func(t *testing.T) {
-		// The attack the T1 evidence exists to refuse: inbound mail whose From
-		// header claims the mailbox owner. It parses as outbound — that is all
-		// direction can mean — but the provider filed it in the inbox and
-		// attested nothing, so the address stays suppressed infrastructure.
-		sync(t, email(autoCreateOwner, "", "blast@sendgrid.net", "forge1@evil.example", ""))
-		if n := countRows(t, e, `
-			SELECT count(*) FROM activity
-			WHERE counterparty_email = 'blast@sendgrid.net' AND counterparty_outbound_attested`); n != 0 {
-			t.Fatal("a forged From:owner message attested correspondence — the gate reads the header, not the provider")
-		}
-		sync(t, emailWithListUnsub("blast@sendgrid.net", "Blast", autoCreateOwner, "forge2@sendgrid.net"))
-		if n := countRows(t, e, `
-			SELECT count(*) FROM person p JOIN person_email pe ON pe.person_id = p.id
-			WHERE pe.email = 'blast@sendgrid.net'`); n != 0 {
-			t.Fatal("a forged From:owner whitelisted an ESP address past T2 suppression")
-		}
-	})
-	t.Run("a conference blast WITHOUT corroboration is a normal counterparty", func(t *testing.T) {
-		// The same prefix subdomain, but no List-Unsubscribe and a human
-		// localpart: not suppressed — a real company can live at event.*.
-		sync(t, email("ada@event.realco.example", "Ada Real", autoCreateOwner, "rc1@event.realco.example", ""))
-		if n := countRows(t, e, `
-			SELECT count(*) FROM person p JOIN person_email pe ON pe.person_id = p.id
-			WHERE pe.email = 'ada@event.realco.example'`); n != 1 {
-			t.Fatal("an uncorroborated prefix sender must create a normal person")
-		}
-	})
-}
-
-// What a captured activity carries: the counterparty identity the correspondence
-// gate reads, and an audit image that is metadata-only.
 func TestCaptureRecordsProvenanceWithoutTheMessageBody(t *testing.T) {
 	env := newCaptureEnv(t)
 	e, sync := env.e, env.sync
